cleanup: docs and module structure
