+++ iroh-dns-server/src/http/pkarr.rs
@@ -7,8 +7,8 @@ use http::{header, StatusCode};
 
 use tracing::info;
 
-use crate::state::{AppState, PacketSource};
 use crate::util::PublicKeyBytes;
+use crate::{state::AppState, store::PacketSource};
 
 use super::error::AppError;
