+++ iroh-dns-server/src/store.rs
@@ -1,98 +1,190 @@
-use std::path::Path;
+//! Pkarr packet store used to resolve DNS queries.
+
+use std::{collections::BTreeMap, num::NonZeroUsize, path::Path, sync::Arc};
 
 use anyhow::Result;
+use hickory_proto::rr::{Name, RecordSet, RecordType, RrKey};
 use iroh_metrics::inc;
+use lru::LruCache;
+use parking_lot::Mutex;
 use pkarr::SignedPacket;
-use redb::{backends::InMemoryBackend, Database, ReadableTable, TableDefinition};
 
-use crate::{metrics::Metrics, util::PublicKeyBytes};
+use crate::{
+    metrics::Metrics,
+    util::{signed_packet_to_hickory_records_without_origin, PublicKeyBytes},
+};
 
-pub type SignedPacketsKey = [u8; 32];
-const SIGNED_PACKETS_TABLE: TableDefinition<&SignedPacketsKey, &[u8]> =
-    TableDefinition::new("signed-packets-1");
+use self::signed_packets::SignedPacketStore;
+
+mod signed_packets;
+
+/// Cache up to 1 million pkarr zones by default
+pub const DEFAULT_CACHE_CAPACITY: usize = 1024 * 1024;
+
+/// Where a new pkarr packet comes from
+pub enum PacketSource {
+    /// Received via HTTPS relay PUT
+    PkarrPublish,
+}
+
+/// A store for pkarr signed packets.
+///
+/// Packets are stored in the persistent [`SignedPacketStore`], and cached on-demand in an in-memory LRU
+/// cache used for resolving DNS queries.
+#[derive(Debug, Clone)]
+pub struct ZoneStore {
+    cache: Arc<Mutex<ZoneCache>>,
+    store: Arc<SignedPacketStore>,
+}
+
+impl ZoneStore {
+    /// Create a persistent store
+    pub fn persistent(path: impl AsRef<Path>) -> Result<Self> {
+        let packet_store = SignedPacketStore::persistent(path)?;
+        Ok(Self::new(packet_store))
+    }
+
+    /// Create an in-memory store.
+    pub fn in_memory() -> Result<Self> {
+        let packet_store = SignedPacketStore::in_memory()?;
+        Ok(Self::new(packet_store))
+    }
+
+    /// Create a new zone store.
+    pub fn new(store: SignedPacketStore) -> Self {
+        let zone_cache = ZoneCache::new(DEFAULT_CACHE_CAPACITY);
+        Self {
+            store: Arc::new(store),
+            cache: Arc::new(Mutex::new(zone_cache)),
+        }
+    }
+
+    /// Resolve a DNS query.
+    // allow unused async: this will be async soon.
+    #[allow(clippy::unused_async)]
+    pub async fn resolve(
+        &self,
+        pubkey: &PublicKeyBytes,
+        name: &Name,
+        record_type: RecordType,
+    ) -> Result<Option<Arc<RecordSet>>> {
+        if let Some(rset) = self.cache.lock().resolve(pubkey, name, record_type) {
+            return Ok(Some(rset));
+        }
+
+        if let Some(packet) = self.store.get(pubkey)? {
+            return self
+                .cache
+                .lock()
+                .insert_and_resolve(&packet, name, record_type);
+        };
+
+        // This would be where mainline discovery could be added.
+
+        Ok(None)
+    }
+
+    /// Get the latest signed packet for a pubkey.
+    // allow unused async: this will be async soon.
+    #[allow(clippy::unused_async)]
+    pub async fn get_signed_packet(&self, pubkey: &PublicKeyBytes) -> Result<Option<SignedPacket>> {
+        self.store.get(pubkey)
+    }
+
+    /// Insert a signed packet into the cache and the store.
+    ///
+    /// Returns whether this produced an update, i.e. whether the packet is the newest for its
+    /// pubkey.
+    // allow unused async: this will be async soon.
+    #[allow(clippy::unused_async)]
+    pub async fn insert(&self, signed_packet: SignedPacket, _source: PacketSource) -> Result<bool> {
+        let pubkey = PublicKeyBytes::from_signed_packet(&signed_packet);
+        if self.store.upsert(signed_packet)? {
+            inc!(Metrics, pkarr_publish_update);
+            self.cache.lock().remove(&pubkey);
+            Ok(true)
+        } else {
+            inc!(Metrics, pkarr_publish_noop);
+            Ok(false)
+        }
+    }
+}
 
 #[derive(Debug)]
-pub struct SignedPacketStore {
-    db: Database,
+struct ZoneCache {
+    cache: LruCache<PublicKeyBytes, CachedZone>,
 }
 
-impl SignedPacketStore {
-    pub fn persistent(path: impl AsRef<Path>) -> Result<Self> {
-        if let Some(parent) = path.as_ref().parent() {
-            std::fs::create_dir_all(parent)?;
-        }
-        let db = Database::builder().create(path)?;
-        Self::open(db)
+impl ZoneCache {
+    fn new(cap: usize) -> Self {
+        let cache = LruCache::new(NonZeroUsize::new(cap).expect("capacity must be larger than 0"));
+        Self { cache }
     }
 
-    pub fn in_memory() -> Result<Self> {
-        let db = Database::builder().create_with_backend(InMemoryBackend::new())?;
-        Self::open(db)
+    fn resolve(
+        &mut self,
+        pubkey: &PublicKeyBytes,
+        name: &Name,
+        record_type: RecordType,
+    ) -> Option<Arc<RecordSet>> {
+        self.cache
+            .get(pubkey)
+            .and_then(|zone| zone.resolve(name, record_type))
     }
 
-    pub fn open(db: Database) -> Result<Self> {
-        let write_tx = db.begin_write()?;
+    fn insert_and_resolve(
+        &mut self,
+        signed_packet: &SignedPacket,
+        name: &Name,
+        record_type: RecordType,
+    ) -> Result<Option<Arc<RecordSet>>> {
+        let pubkey = PublicKeyBytes::from_signed_packet(signed_packet);
+        self.insert(signed_packet)?;
+        Ok(self.resolve(&pubkey, name, record_type))
+    }
+
+    fn insert(&mut self, signed_packet: &SignedPacket) -> Result<()> {
+        let pubkey = PublicKeyBytes::from_signed_packet(signed_packet);
+        if self
+            .cache
+            .peek(&pubkey)
+            .map(|old| old.is_newer_than(signed_packet))
+            .unwrap_or(false)
         {
-            let _table = write_tx.open_table(SIGNED_PACKETS_TABLE)?;
+            return Ok(());
         }
-        write_tx.commit()?;
-        Ok(Self { db })
+        self.cache
+            .put(pubkey, CachedZone::from_signed_packet(signed_packet)?);
+        Ok(())
     }
 
-    pub fn upsert(&self, packet: SignedPacket) -> Result<bool> {
-        let key = PublicKeyBytes::from_signed_packet(&packet);
-        let tx = self.db.begin_write()?;
-        let mut replaced = false;
-        {
-            let mut table = tx.open_table(SIGNED_PACKETS_TABLE)?;
-            if let Some(existing) = get_packet(&table, &key)? {
-                if existing.more_recent_than(&packet) {
-                    return Ok(false);
-                } else {
-                    replaced = true;
-                }
-            }
-            let value = packet.as_bytes();
-            table.insert(key.as_bytes(), &value[..])?;
-        }
-        tx.commit()?;
-        if replaced {
-            inc!(Metrics, store_packets_updated);
-        } else {
-            inc!(Metrics, store_packets_inserted);
-        }
-        Ok(true)
-    }
-
-    pub fn get(&self, key: &PublicKeyBytes) -> Result<Option<SignedPacket>> {
-        let tx = self.db.begin_read()?;
-        let table = tx.open_table(SIGNED_PACKETS_TABLE)?;
-        get_packet(&table, key)
-    }
-
-    pub fn remove(&self, key: &PublicKeyBytes) -> Result<bool> {
-        let tx = self.db.begin_write()?;
-        let updated = {
-            let mut table = tx.open_table(SIGNED_PACKETS_TABLE)?;
-            let did_remove = table.remove(key.as_bytes())?.is_some();
-            #[allow(clippy::let_and_return)]
-            did_remove
-        };
-        tx.commit()?;
-        if updated {
-            inc!(Metrics, store_packets_removed)
-        }
-        Ok(updated)
+    fn remove(&mut self, pubkey: &PublicKeyBytes) {
+        self.cache.pop(pubkey);
     }
 }
 
-fn get_packet(
-    table: &impl ReadableTable<&'static SignedPacketsKey, &'static [u8]>,
-    key: &PublicKeyBytes,
-) -> Result<Option<SignedPacket>> {
-    let Some(row) = table.get(key.as_ref())? else {
-        return Ok(None);
-    };
-    let packet = SignedPacket::from_bytes(row.value().to_vec().into(), false)?;
-    Ok(Some(packet))
+#[derive(Debug)]
+struct CachedZone {
+    timestamp: u64,
+    records: BTreeMap<RrKey, Arc<RecordSet>>,
+}
+
+impl CachedZone {
+    fn from_signed_packet(signed_packet: &SignedPacket) -> Result<Self> {
+        let (_label, records) =
+            signed_packet_to_hickory_records_without_origin(signed_packet, |_| true)?;
+        Ok(Self {
+            records,
+            timestamp: *signed_packet.timestamp(),
+        })
+    }
+
+    fn is_newer_than(&self, signed_packet: &SignedPacket) -> bool {
+        self.timestamp > *signed_packet.timestamp()
+    }
+
+    fn resolve(&self, name: &Name, record_type: RecordType) -> Option<Arc<RecordSet>> {
+        let key = RrKey::new(name.into(), record_type);
+        self.records.get(&key).cloned()
+    }
 }
